requirements.txt
deepeval
google-generativeai
python-dotenv